Rework modules structure
