--- app/PscPackage/Types.hs
@@ -1,18 +1,0 @@
-{-# LANGUAGE TemplateHaskell #-}
-
-module PscPackage.Types where
-
-import           Data.Aeson
-import           Data.Aeson.TH
-import           Data.Text     (Text)
-import           GHC.Generics
-
-data PscPackage = PscPackage
-  { name    :: Text
-  , set     :: Text
-  , source  :: Text
-  , depends :: [Text]
-  }
-  deriving (Show, Generic)
-
-$(deriveJSON defaultOptions ''PscPackage)
